--- src/main/kotlin/com/deepromeet/atcha/transit/domain/BusRouteWithOperationInfo.kt
@@ -1,6 +1,0 @@
-package com.deepromeet.atcha.transit.domain
-
-data class BusRouteWithOperationInfo(
-    val route: BusRoute,
-    val operationInfo: BusRouteOperationInfo
-)
