+++ src/main/kotlin/com/deepromeet/atcha/transit/infrastructure/client/public/PublicIncheonRouteInfoClient.kt
@@ -1,10 +1,13 @@
 package com.deepromeet.atcha.transit.infrastructure.client.public
 
+import com.deepromeet.atcha.location.domain.CoordinateTransformer
 import com.deepromeet.atcha.transit.domain.BusRealTimeArrival
 import com.deepromeet.atcha.transit.domain.BusRoute
 import com.deepromeet.atcha.transit.domain.BusRouteInfo
 import com.deepromeet.atcha.transit.domain.BusRouteInfoClient
+import com.deepromeet.atcha.transit.domain.BusRouteInfoClient.Companion.NON_STOP_STATION_NAME
 import com.deepromeet.atcha.transit.domain.BusRouteOperationInfo
+import com.deepromeet.atcha.transit.domain.BusRouteStationList
 import com.deepromeet.atcha.transit.domain.BusSchedule
 import com.deepromeet.atcha.transit.exception.TransitError
 import com.deepromeet.atcha.transit.exception.TransitException
@@ -14,6 +17,7 @@ import org.springframework.stereotype.Component
 @Component
 class PublicIncheonRouteInfoClient(
     private val incheonBusRouteInfoFeignClient: PublicIncheonBusRouteInfoFeignClient,
+    private val coordinateTransformer: CoordinateTransformer,
     @Value("\${open-api.api.service-key}")
     private val serviceKey: String,
     @Value("\${open-api.api.spare-key}")
@@ -83,6 +87,44 @@ class PublicIncheonRouteInfoClient(
                     )
             },
             errorMessage = "인천시 버스 노선-${route.name}-${route.id.value}의 운영 정보를 가져오는데 실패했습니다."
+        )
+    }
+
+    override fun getStationList(route: BusRoute): BusRouteStationList {
+        return ApiClientUtils.callApiWithRetry(
+            primaryKey = serviceKey,
+            spareKey = spareKey,
+            realLastKey = realLastKey,
+            apiCall = { key ->
+                incheonBusRouteInfoFeignClient.getBusRouteSectionList(key, route.id.value)
+            },
+            isLimitExceeded = { response -> ApiClientUtils.isServiceResultApiLimitExceeded(response) },
+            processResult = { response ->
+                val turnPoint = response.msgBody.itemList?.first { it.directionCode == 1 }
+
+                val routeStations =
+                    response.msgBody.itemList
+                        ?.filter { station ->
+                            NON_STOP_STATION_NAME.none { keyword -> station.stationName.contains(keyword) }
+                        }
+                        ?.map {
+                            it.toBusRouteStation(
+                                route,
+                                turnPoint?.stationSequence,
+                                coordinateTransformer.transformToWGS84(it.positionX, it.positionY)
+                            )
+                        }
+                        ?: throw TransitException.of(
+                            TransitError.NOT_FOUND_BUS_STATION,
+                            "인천시 버스 노선-${route.name}-${route.id.value}의 경유 정류소 정보를 찾을 수 없습니다."
+                        )
+
+                BusRouteStationList(
+                    routeStations,
+                    turnPoint?.stationSequence
+                )
+            },
+            errorMessage = "인천시 버스 노선 스케줄 정보를 가져오는데 실패했습니다."
         )
     }
