+++ src/main/kotlin/com/deepromeet/atcha/transit/infrastructure/client/public/PublicSeoulBusRouteInfoClient.kt
@@ -4,19 +4,22 @@ import com.deepromeet.atcha.transit.domain.BusRealTimeArrival
 import com.deepromeet.atcha.transit.domain.BusRoute
 import com.deepromeet.atcha.transit.domain.BusRouteInfo
 import com.deepromeet.atcha.transit.domain.BusRouteInfoClient
+import com.deepromeet.atcha.transit.domain.BusRouteInfoClient.Companion.NON_STOP_STATION_NAME
 import com.deepromeet.atcha.transit.domain.BusRouteOperationInfo
+import com.deepromeet.atcha.transit.domain.BusRouteStationList
 import com.deepromeet.atcha.transit.domain.BusSchedule
 import com.deepromeet.atcha.transit.exception.TransitError
 import com.deepromeet.atcha.transit.exception.TransitException
-import io.github.oshai.kotlinlogging.KotlinLogging
+import com.deepromeet.atcha.transit.infrastructure.client.public.ApiClientUtils.isServiceResultApiLimitExceeded
 import org.springframework.beans.factory.annotation.Value
 import org.springframework.stereotype.Component
-
-private val log = KotlinLogging.logger {}
+import kotlin.collections.filter
+import kotlin.collections.first
 
 @Component
 class PublicSeoulBusRouteInfoClient(
     private val publicSeoulBusArrivalInfoFeignClient: PublicSeoulBusArrivalInfoFeignClient,
+    private val publicBusRouteClient: PublicSeoulBusRouteInfoFeignClient,
     private val seoulBusOperationFeignClient: SeoulBusOperationFeignClient,
     @Value("\${open-api.api.service-key}")
     private val serviceKey: String,
@@ -31,7 +34,7 @@ class PublicSeoulBusRouteInfoClient(
             spareKey = spareKey,
             realLastKey = realLastKey,
             apiCall = { key -> publicSeoulBusArrivalInfoFeignClient.getBusRouteList(key, routeName) },
-            isLimitExceeded = { response -> ApiClientUtils.isServiceResultApiLimitExceeded(response) },
+            isLimitExceeded = { response -> isServiceResultApiLimitExceeded(response) },
             processResult = { response ->
                 val routes = (
                     response.msgBody.itemList
@@ -80,6 +83,35 @@ class PublicSeoulBusRouteInfoClient(
                 TransitError.NOT_FOUND_BUS_OPERATION_INFO,
                 "서울시 버스 노선 '${route.id.value}'의 운영 정보를 찾을 수 없습니다."
             )
+    }
+
+    override fun getStationList(route: BusRoute): BusRouteStationList {
+        return ApiClientUtils.callApiWithRetry(
+            primaryKey = serviceKey,
+            spareKey = spareKey,
+            realLastKey = realLastKey,
+            apiCall = { key -> publicBusRouteClient.getStationsByRoute(route.id.value, key) },
+            isLimitExceeded = { response -> isServiceResultApiLimitExceeded(response) },
+            processResult = { response ->
+                val responses = response.msgBody.itemList
+                if (responses == null) {
+                    throw TransitException.of(
+                        TransitError.BUS_ROUTE_STATION_LIST_FETCH_FAILED,
+                        "버스 노선 '${route.id.value}'의 경유 정류소 정보를 찾을 수 없습니다."
+                    )
+                } else {
+                    val turnPoint = responses.first { it.transYn == "Y" }.seq.toInt()
+                    val busRouteStations =
+                        responses
+                            .filter { station ->
+                                NON_STOP_STATION_NAME.none { keyword -> station.stationNm.contains(keyword) }
+                            }
+                            .map { it.toBusRouteStation(turnPoint) }
+                    BusRouteStationList(busRouteStations, turnPoint)
+                }
+            },
+            errorMessage = "서울시 버스 노선 경유 정류소를 가져오는데 실패했습니다."
+        )
     }
 
     override fun getBusRealTimeInfo(routeInfo: BusRouteInfo): BusRealTimeArrival {
